+++ packages/plugin-expo-device/device.js
@@ -2,6 +2,10 @@ const Device = require('expo-device')
 const Constants = require('expo-constants').default
 const { Dimensions, Platform } = require('react-native')
 const rnVersion = require('react-native/package.json').version
+const cuid = require('@bugsnag/cuid')
+const SecureStore = require('expo-secure-store')
+
+const DEVICE_ID_KEY = 'bugsnag-anonymous-id'
 
 module.exports = {
   load: client => {
@@ -21,8 +25,19 @@ module.exports = {
     // get the initial orientation
     updateOrientation()
 
+    const storeOptions = {
+      requireAuthentication: false,
+      keychainAccessible: SecureStore.ALWAYS_THIS_DEVICE_ONLY
+    }
+
+    let deviceId = SecureStore.getItem(DEVICE_ID_KEY, storeOptions)
+    if (!deviceId || !cuid.isCuid(deviceId)) {
+      deviceId = cuid()
+      SecureStore.setItem(DEVICE_ID_KEY, deviceId, storeOptions)
+    }
+
     const device = {
-      id: Constants.installationId,
+      id: deviceId,
       manufacturer: Device.manufacturer,
       model: Device.modelName,
       modelNumber: Device.modelId || undefined,